+++ src/main/java/com/doublew2w/sbs/mybatis/executor/ExecutionPlaceholder.java
@@ -1,0 +1,11 @@
+package com.doublew2w.sbs.mybatis.executor;
+
+/**
+ * 占位符
+ * @author: DoubleW2w
+ * @date: 2024/9/24 4:03
+ * @project: sbs-mybatis
+ */
+public enum ExecutionPlaceholder {
+  EXECUTION_PLACEHOLDER
+}