+++ src/main/java/com/doublew2w/sbs/mybatis/executor/Executor.java
@@ -1,5 +1,6 @@
 package com.doublew2w.sbs.mybatis.executor;
 
+import com.doublew2w.sbs.mybatis.cache.CacheKey;
 import com.doublew2w.sbs.mybatis.mapping.BoundSql;
 import com.doublew2w.sbs.mybatis.mapping.MappedStatement;
 import com.doublew2w.sbs.mybatis.session.ResultHandler;
@@ -18,14 +19,17 @@ import java.util.List;
 public interface Executor {
   ResultHandler NO_RESULT_HANDLER = null;
 
-  /** 查询 */
+  // 查询，含缓存
   <E> List<E> query(
       MappedStatement ms,
       Object parameter,
       RowBounds rowBounds,
       ResultHandler resultHandler,
-      BoundSql boundSql);
+      CacheKey key,
+      BoundSql boundSql)
+      throws SQLException;
 
+  // 查询
   <E> List<E> query(
       MappedStatement ms, Object parameter, RowBounds rowBounds, ResultHandler resultHandler)
       throws SQLException;
@@ -40,4 +44,19 @@ public interface Executor {
   void rollback(boolean required) throws SQLException;
 
   void close(boolean forceRollback);
+
+  /** 清理Session缓存 */
+  void clearLocalCache();
+
+  /**
+   * 创建缓存 Key
+   *
+   * @param ms 映射语句
+   * @param parameterObject 映射语句参数
+   * @param rowBounds 分页记录
+   * @param boundSql 绑定SQL
+   * @return 缓存Key
+   */
+  CacheKey createCacheKey(
+      MappedStatement ms, Object parameterObject, RowBounds rowBounds, BoundSql boundSql);
 }